Refactored FacebookJob
Now all facebook jobs are in 1 file. We (currently) have 3 classes:
* FacebookAddJob
* FacebookDelteJob
* FacebookGetJob

And they all have FacebookJob as a parent.

They are all in 1 file since the classes are small and do not contain to much
logic. All the other classes depending on one of the above classes have ben
changed.